Completing 8601 Algorithm Check
Have enough confidence in the new algorthm that needs to be generated
every year. Now, trying to figure out how to update on a daily basis.

Testing out two methods:
1. by wakeup task (not certain when this is triggered)
2. when plugged into charger (at least weekly, which is also a nice
   sanity check on Mondays)

Made some styling updates as well. Critical information is in bold, with
nont-critical in light.

Long term question: should battery percentage be highlighted? Will I
charge it based on how much battery is on the device? No... therefore,
does it need to be displayed?