Updated for 2024 and heartrate
Weeknum was re-created with google sheets to ensure that it calculates
correctly for next year. This year, 30 and 31 DEC will look funny (will
return 53)

Heartrate now displays because teammate asked for a display. But he
might have a Versa 4, so may not be able to take this.

Going forward, newest Fitbits do not support custom watch faces
(google-ized). So, probably only 3 years left of this useful repository.